requirements.txt
fastapi
uvicorn
openai
langchain
langchain-community
langchain-openai
langchain-pinecone
pinecone-client
tiktoken
pypdf
